Fix broken builds
