+++ agent/app/agent_unspecified.go
@@ -1,6 +1,6 @@
 // +build !linux,!windows
 
-// Copyright 2014-2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// Copyright 2017 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License"). You may
 // not use this file except in compliance with the License. A copy of the
@@ -16,92 +16,36 @@
 package app
 
 import (
-	"strings"
+	"errors"
 
-	"github.com/aws/amazon-ecs-agent/agent/config"
-	"github.com/aws/amazon-ecs-agent/agent/dockerclient"
-	"github.com/aws/amazon-ecs-agent/agent/dockerclient/dockerapi"
+	"github.com/aws/amazon-ecs-agent/agent/credentials"
 	"github.com/aws/amazon-ecs-agent/agent/ecs_client/model/ecs"
-	"github.com/aws/amazon-ecs-agent/agent/taskresource/volume"
+	"github.com/aws/amazon-ecs-agent/agent/engine"
+	"github.com/aws/amazon-ecs-agent/agent/engine/dockerstate"
 	"github.com/cihub/seelog"
 )
 
-func (agent *ecsAgent) appendVolumeDriverCapabilities(capabilities []*ecs.Attribute) []*ecs.Attribute {
-	// "local" is default docker driver
-	capabilities = appendNameOnlyAttribute(capabilities, attributePrefix+capabilityDockerPluginInfix+volume.DockerLocalVolumeDriver)
-
-	// for non-standardized plugins, call docker pkg's plugins.Scan()
-	nonStandardizedPlugins, err := agent.mobyPlugins.Scan()
-	if err != nil {
-		seelog.Warnf("Scanning plugins failed: %v", err)
-		// do not return yet, we need the list of plugins below. range handles nil slice.
-	}
-
-	for _, pluginName := range nonStandardizedPlugins {
-		// Replace the ':' to '.' in the plugin name for attributes
-		capabilities = appendNameOnlyAttribute(capabilities,
-			attributePrefix+capabilityDockerPluginInfix+strings.Replace(pluginName, config.DockerTagSeparator, attributeSeparator, -1))
-	}
-
-	// for standardized plugins, call docker's plugin ls API
-	pluginEnabled := true
-	volumeDriverType := []string{dockerapi.VolumeDriverType}
-	standardizedPlugins, err := agent.dockerClient.ListPluginsWithFilters(agent.ctx, pluginEnabled, volumeDriverType, dockerclient.ListPluginsTimeout)
-	if err != nil {
-		seelog.Warnf("Listing plugins with filters enabled=%t, capabilities=%v failed: %v", pluginEnabled, volumeDriverType, err)
-		return capabilities
-	}
-
-	// For plugin with default tag latest, register two attributes with and without the latest tag
-	// as the tag is optional and can be added by docker or customer
-	for _, pluginName := range standardizedPlugins {
-		names := strings.Split(pluginName, config.DockerTagSeparator)
-		if len(names) > 1 && names[len(names)-1] == config.DefaultDockerTag {
-			capabilities = appendNameOnlyAttribute(capabilities, attributePrefix+capabilityDockerPluginInfix+strings.Join(names[:len(names)-1], attributeSeparator))
-		}
-
-		capabilities = appendNameOnlyAttribute(capabilities,
-			attributePrefix+capabilityDockerPluginInfix+strings.Replace(pluginName, config.DockerTagSeparator, attributeSeparator, -1))
-	}
-	return capabilities
+func (agent *ecsAgent) initializeTaskENIDependencies(state dockerstate.TaskEngineState, taskEngine engine.TaskEngine) (error, bool) {
+	return errors.New("unsupported platform"), true
 }
 
-func (agent *ecsAgent) appendNvidiaDriverVersionAttribute(capabilities []*ecs.Attribute) []*ecs.Attribute {
-	return capabilities
+// startWindowsService is not supported on non windows platforms
+func (agent *ecsAgent) startWindowsService() int {
+	seelog.Error("Windows Services are not supported on unspecified platforms")
+	return 1
 }
 
-func (agent *ecsAgent) appendENITrunkingCapabilities(capabilities []*ecs.Attribute) []*ecs.Attribute {
-	return capabilities
+func (agent *ecsAgent) initializeResourceFields(credentialsManager credentials.Manager) {
 }
 
-func (agent *ecsAgent) appendPIDAndIPCNamespaceSharingCapabilities(capabilities []*ecs.Attribute) []*ecs.Attribute {
-	return capabilities
+func (agent *ecsAgent) cgroupInit() error {
+	return nil
 }
 
-func (agent *ecsAgent) appendAppMeshCapabilities(capabilities []*ecs.Attribute) []*ecs.Attribute {
-	return capabilities
+func (agent *ecsAgent) initializeGPUManager() error {
+	return nil
 }
 
-func (agent *ecsAgent) appendTaskEIACapabilities(capabilities []*ecs.Attribute) []*ecs.Attribute {
-	return capabilities
-}
-
-func (agent *ecsAgent) appendFirelensFluentdCapabilities(capabilities []*ecs.Attribute) []*ecs.Attribute {
-	return capabilities
-}
-
-func (agent *ecsAgent) appendFirelensFluentbitCapabilities(capabilities []*ecs.Attribute) []*ecs.Attribute {
-	return capabilities
-}
-
-func (agent *ecsAgent) appendFirelensLoggingDriverCapabilities(capabilities []*ecs.Attribute) []*ecs.Attribute {
-	return capabilities
-}
-
-func (agent *ecsAgent) appendFirelensConfigCapabilities(capabilities []*ecs.Attribute) []*ecs.Attribute {
-	return capabilities
-}
-
-func (agent *ecsAgent) appendGMSACapabilities(capabilities []*ecs.Attribute) []*ecs.Attribute {
-	return capabilities
+func (agent *ecsAgent) getPlatformDevices() []*ecs.PlatformDevice {
+	return nil
 }
